% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GCclassifier.R
\docType{package}
\name{GCclassifier}
\alias{GCclassifier}
\title{GCclassifier: an R package for prediction of gene
expression-based molecular subtype of gastric cancer}
\description{
Gastric cancer (GC) can be divided into gene
expression-based and biologically distinct molecular subtypes.
This package is to built to label gastric cancer samples based
on their log2 scaled gene expression profiles (GEP)
}
\references{
Oh, S.C., Sohn, B.H., Cheong, JH. et al. Clinical and genomic landscape of gastric
cancer with a mesenchymal phenotype. Nat Commun 9, 1777 (2018).
https://doi.org/10.1038/s41467-018-04179-8.

Cristescu, R., Lee, J., Nebozhyn, M. et al. Molecular analysis of gastric cancer
identifies subtypes associated with distinct clinical outcomes. Nat Med 21, 449–456 (2015).
https://doi.org/10.1038/nm.3850.

The Cancer Genome Atlas Research Network. Comprehensive molecular characterization of
gastric adenocarcinoma. Nature 513, 202–209 (2014).
https://doi.org/10.1038/nature13480
}
