% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_molecular_subtype.R
\name{get_molecular_subtype}
\alias{get_molecular_subtype}
\title{Gastric cancer molecular subtypes prediction}
\usage{
get_molecular_subtype(
  Expr = NULL,
  method = "EMP",
  idType = "SYMBOL",
  minPosterior = 0.5,
  maxp = NULL,
  verbose = TRUE
)
}
\arguments{
\item{Expr}{a dataframe or matrix with log2 scaled gene expression profile data,
with samples in column and genes in rows, gene names should not be empty.}

\item{method}{subtyping classification model, should be either "ACRG", "EMP" or "TCGA". By default is "EMP".}

\item{idType}{a string which indicates which type of gene ids used in the rownames of GEP,
should be one of "SYMBOL","ENSEMBL","ENTREZID" and "REFSEQ". By default is "SYMBOL".}

\item{minPosterior}{minimal posterior probability to classify a sample in TCGA subtyping system. By default is 0.5.}

\item{maxp}{the maxp parameter used in \code{\link[impute]{impute.knn}} function,
it is optional.}

\item{verbose}{a single logical value specifying to display detailed messages (when verbose=TRUE)
or not (when verbose=FALSE).}
}
\value{
a dataframe with sample names and predicted molecular subtype labels. If the method
is one of "EMP" and "TCGA", predicted probabilities of each subtype (with corresponding
subtype column names) will be included.
}
\description{
Predict the molecular subtypes of gastric cancer samples,
based on log2 scaled gene expression profile (GEP).
}
\examples{
data("GSE62254")
result <- get_molecular_subtype(Expr = GSE62254, method = "ACRG", idType = "SYMBOL")
}
\references{
Oh, S.C., Sohn, B.H., Cheong, JH. et al. Clinical and genomic landscape of gastric
cancer with a mesenchymal phenotype. Nat Commun 9, 1777 (2018).
https://doi.org/10.1038/s41467-018-04179-8.

Cristescu, R., Lee, J., Nebozhyn, M. et al. Molecular analysis of gastric cancer
identifies subtypes associated with distinct clinical outcomes. Nat Med 21, 449–456 (2015).
https://doi.org/10.1038/nm.3850.

The Cancer Genome Atlas Research Network. Comprehensive molecular characterization of
gastric adenocarcinoma. Nature 513, 202–209 (2014).
https://doi.org/10.1038/nature13480
}
