% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_molecular_subtype_interface.R
\name{get_molecular_subtype_interface}
\alias{get_molecular_subtype_interface}
\title{The shiny interface of gastric cancer molecular subtype prediction}
\usage{
get_molecular_subtype_interface()
}
\description{
This function runs the shiny app for subtype prediction of gastric cancer.
}
\examples{
get_molecular_subtype_interface()
}
