% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_signature.R
\name{get_signature}
\alias{get_signature}
\title{Extract signature genes in symbol/entrez ids from a specific gastric cancer molecular subtype classifier}
\usage{
get_signature(method = "EMP")
}
\arguments{
\item{method}{Molecular subtype for which features should be reported. This can be a string of 'EMP','TCGA' and 'ACRG' (default EMP)}
}
\value{
A data frame with signature genes
}
\description{
Extract signature genes in symbol/entrez ids from a specific gastric cancer molecular subtype classifier
}
\examples{
get_signature(method = 'EMP')
}
