% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE26901}
\alias{GSE26901}
\title{Gene expression profile of GEO dataset GSE26901}
\format{
A gene expression data frame with 19236 gene symbols
and 109 samples:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE26901}
}
\usage{
GSE26901
}
\description{
A dataset containing the gene expression profile of 109
gastric cancer samples.
}
\keyword{datasets}
