% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE26901_subtype}
\alias{GSE26901_subtype}
\title{Phenotype and subtype information of GEO dataset GSE26901}
\format{
A dataframe with 109 observations and 12 variables:
\describe{
 \item{Patients_ID}{patient id number}
 \item{Subgroup}{subtype information in EMP subtyping system}
 \item{Sex}{M: male; F: demale}
 \item{Age}{age information}
 \item{AJCC.stage}{AJCC TNM stage}
 \item{M.stage}{metastasis stage (0/1)}
 \item{Death (1=yes, 0=no)}{survial condition}
 \item{OS.m}{overall survival (OS), in months}
 \item{Recurrence (1=yes, 0=no)}{recurrence information}
 \item{RFS.m}{recurrence free survival (RFS), in months}
 \item{Adjuvant.chemo}{receive adjuvant chemo}
 \item{geo_accession}{GEO accession number}
}
}
\source{
\url{https://www.nature.com/articles/s41467-018-04179-8}
}
\usage{
GSE26901_subtype
}
\description{
A dataframe containing the subtype and survival information of 109
gastric cancer samples.
}
\keyword{datasets}
