% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE62254}
\alias{GSE62254}
\title{Gene expression profile of GEO dataset GSE62254}
\format{
A gene expression data frame with 1055 gene symbols
and 300 samples:
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE62254}
}
\usage{
GSE62254
}
\description{
A dataset containing the gene expression profile of 300
gastric cancer samples.
}
\keyword{datasets}
