% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE62254_subtype}
\alias{GSE62254_subtype}
\title{Phenotype and subtype information of GEO dataset GSE62254}
\format{
A dataframe with 300 observations and 7 variables:
\describe{
 \item{GEO_ID}{GEO id number}
 \item{Subgroup}{subtype information in EMP subtyping system}
 \item{Recur}{recurrence information, 1 means with recurrence}
 \item{DFS.m}{disease free survival (DFS), in months}
 \item{Death}{survial condition, 1 means death}
 \item{OS.m}{overall survival (OS), in months}
 \item{ACRG.sub}{subtype information in ACRG subtyping system}
}
}
\source{
\url{https://www.nature.com/articles/nm.3850}
}
\usage{
GSE62254_subtype
}
\description{
A dataframe containing the subtype and survival information of 300
gastric cancer samples.
}
\keyword{datasets}
