% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log2IfNeeded.R
\name{log2IfNeeded}
\alias{log2IfNeeded}
\title{Log2 scale gene expression if needed}
\usage{
log2IfNeeded(df = NULL)
}
\arguments{
\item{df}{a data frame/matrix of gene expression that will be log2 transformed (default: NULL)}
}
\value{
log2 scaled gene expression in data frame or matrix format
}
\description{
Log2 scale gene expression if needed
}
