% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{toLog2}
\alias{toLog2}
\title{To check whether the GEP should perform log2 transformation}
\usage{
toLog2(df)
}
\arguments{
\item{df}{a dataframe of GEP}
}
\value{
a logical value (TRUE/FALSE)
}
\description{
To check whether the GEP should perform log2 transformation
}
