% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_molecular_subtype_interface.R
\name{get_molecular_subtype_interface}
\alias{get_molecular_subtype_interface}
\title{The shiny interface of gastric cancer molecular subtypes prediction}
\usage{
get_molecular_subtype_interface()
}
\description{
This function runs the shiny app for subtypes prediction of gastric cancer. The maximum upload file size is 30MB.
}
\examples{
get_molecular_subtype_interface()
}
