% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probe_to_symbol.R
\name{probe_to_symbol}
\alias{probe_to_symbol}
\title{Aggregate gene expression profile in probe level to symbol level}
\usage{
probe_to_symbol(Expr, annotation, aggregateMethod = "max")
}
\arguments{
\item{Expr}{a dataframe or matrix with log2 scaled gene expression profile, samples in columns,
probes in rows, rownames corresponding to the probe in annotation file.}

\item{annotation}{probe annotation file, must have "probe" and "symbol" in colnames,
"probe" corresponding to rownames in Expr and "sample" corresponding to the
specified gene identifier.}

\item{aggregateMethod}{the method to calculate expression values if multiple probes
map to the same symbol. The default is max.}
}
\value{
a dataframe with log2 scaled gene expression profile in symbol level.
}
\description{
This function could help to convert the gene expression from probe level to
symbol (SYMBOL, ENSEMBL, ENTREZID) level.
}
\examples{
\dontrun{
expr <- data.frame('v1'=runif(5), 'v2'=runif(5))
rownames(expr) <- paste0('xx', 1:5)
anno <- data.frame('probe'=paste0('xx', 1:5), 'symbol'=c(letters[1:4], 'd'))
probe_to_symbol(Expr = expr, annotation = anno, aggregateMethod = 'max')
}
}
